ppt-service/requirements.txt
fastapi
uvicorn
python-pptx
pydantic
